Adding version specific tests

+++ scripts/ps_keyring_plugins_test/ps_keyring_plugins_test.sh
@@ -6,6 +6,16 @@ if [ -f /etc/mysql/my.cnf ]; then
 else
   MYCNF="/etc/my.cnf"
 fi
+ 
+if [ -z "$1" ]; then
+  echo "This script needs parameter ps57|ps80"
+  exit 1
+elif [ "$1" != "ps57" -a "$1" != "ps80" ]; then
+  echo "Version not recognized!"
+  exit 1
+else
+  VERSION="$1"
+fi
 
 service mysql stop
 sleep 10
@@ -27,21 +37,23 @@ mysql -e "CREATE FUNCTION keyring_key_remove returns integer SONAME 'keyring_udf
 mysql -e "CREATE FUNCTION keyring_key_generate returns integer SONAME 'keyring_udf.so';"
 mysql -e "CREATE FUNCTION keyring_key_store returns integer SONAME 'keyring_udf.so';"
 
-# keyring_file plugin test
-#mysql -e "INSTALL PLUGIN keyring_file SONAME 'keyring_file.so';"
-mysql -e "CREATE DATABASE IF NOT EXISTS test;"
-mysql --database=test -e "CREATE TABLESPACE ts1 ADD DATAFILE 'ts1.ibd' ENCRYPTION='Y';"
-mysql --database=test -e "CREATE TABLE keyring_file_test (a INT PRIMARY KEY) TABLESPACE ts1 ENCRYPTION='Y';"
-mysql --database=test -e "INSERT INTO keyring_file_test VALUES (1),(2),(3);"
-mysql --database=test -e "ALTER INSTANCE ROTATE INNODB MASTER KEY;"
-result=$(mysql --database=test -N -s -e "CHECKSUM TABLE keyring_file_test;" | awk -F' ' '{print $2}')
-if [ "${result}" != "2050879373" ]; then
-  echo "Data in keyring_file_test table is corrupted!"
-  exit 1
+if ["$VERSION" == "ps57" ]; then
+  # keyring_file plugin test
+  #mysql -e "INSTALL PLUGIN keyring_file SONAME 'keyring_file.so';"
+  mysql -e "CREATE DATABASE IF NOT EXISTS test;"
+  mysql --database=test -e "CREATE TABLESPACE ts1 ADD DATAFILE 'ts1.ibd' ENCRYPTION='Y';"
+  mysql --database=test -e "CREATE TABLE keyring_file_test (a INT PRIMARY KEY) TABLESPACE ts1 ENCRYPTION='Y';"
+  mysql --database=test -e "INSERT INTO keyring_file_test VALUES (1),(2),(3);"
+  mysql --database=test -e "ALTER INSTANCE ROTATE INNODB MASTER KEY;"
+  result=$(mysql --database=test -N -s -e "CHECKSUM TABLE keyring_file_test;" | awk -F' ' '{print $2}')
+  if [ "${result}" != "2050879373" ]; then
+    echo "Data in keyring_file_test table is corrupted!"
+    exit 1
+  fi
+  mysql --database=test -e "DROP TABLE keyring_file_test;"
+  mysql --database=test -e "DROP TABLESPACE ts1;"
+  mysql -e "UNINSTALL PLUGIN keyring_file;"
 fi
-mysql --database=test -e "DROP TABLE keyring_file_test;"
-mysql --database=test -e "DROP TABLESPACE ts1;"
-mysql -e "UNINSTALL PLUGIN keyring_file;"
 
 # service restart so that plugins don't mess with eachother
 service mysql stop
